Commit Recipes API working, all data is present but the view is kinda ugly, need to fix that

+++ app/src/main/java/com/example/inhuis/ui/ingredients/Ingredient.kt
@@ -5,7 +5,7 @@ data class Ingredient(
     var id: Int,
     var name: String,
     var imageURL: String,
-    var amount: Int
+    var amount: Double
 
 ) {
